refactor: 무한 스크롤 훅을 메세지 엔티티로 이동 (#94)
