+++ src/widgets/chat-window/ui/MessageList.tsx
@@ -10,7 +10,7 @@ import { ChevronDownIcon } from 'lucide-react'
 import EmptyState from '@/shared/ui/EmptyState'
 import { cn } from '@/shared/lib/cn'
 import { useChatMessageList } from '@/entities/message/api/queries'
-import useInfiniteScroll from '@/features/chat-message-scroll/lib/useInfiniteScroll'
+import useInfiniteScroll from '@/entities/message/model/useInfiniteScroll'
 import useMessageSubscribe from '@/features/chat-message-subscribe/model/useMessageSubscribe'
 import { useSessionStore } from '@/entities/session/store/session-store'
 import { useTokenStore } from '@/entities/session/store/token-store'
